javascript stuff

+++ 5-javascript/101-function.js
@@ -1,4 +1,4 @@
-function{
+function double(){
   console.log("Hello world");
 }
